Introducing major changes
Metaprogramming was used in order to dynamically create methods and
remove the hardcoded actions and urls. A call is made to
https://foaas.com/operations in order to get all the operations as well
as their paths. Using metaprogramming, this library is always up-to-date
regardless of any possible change that is made to the actions on the
foaas framework.

The setFuckingAddress() method was added in order for the user to be
able to declare a custom address in case he has a personal setup of
foaas framework. The default address that is used is https://foaas.com
because it supports both IPv4 and IPv6.

The FuckingOperations() class is created in order to get and manipulate
all the results from the /operations path provided by foaas.

The CLI support was removed in order for a user to use this library
within his code and not as a standalone library.