Project updated
The LICENCE of this project has been changed. The code that is used from
the old project, has its own licence. The structure has been changed in
order to distinguish the old code from the new one. Finally, the test
files has been changed in order to be able to run with the new project
structure.