Patch for the random() method
Closes #1
A patch was introduced that resolves this bug issue. Instead of using
the "field" parameter from the json, a temporary dictionary that is a
deep copy of the kwargs dictionary, was introduced. From this temporary
dictionary the pop() method is used in order to remove the "shoutcloud"
and the "i18n" parameters, if exist, thus the random() method is able to
executed without throwing an exception if these arguments are present.

"i18n" parameter changed into "lang" parameter (more friendly).